requirements.txt
# Essential only
langchain
vllm
torch
transformers
chromadb
sentence-transformers
PyMuPDF
fastapi
uvicorn
python-multipart
Pillow
numpy
aiofiles